+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- User settings for caching (pokeapi.CacheSettings)
+  - CustomExpire: set a custom cache expiration (in minutes)
+  - UseCache: turn caching on/off
+
 ## [1.3.2] / 2020.04.03
 ### Added
 - Current version to README.